Renamed all helper files to follow camelCase convention
