+++ client/app/bundles/course/assessment/components/Material.jsx
@@ -5,7 +5,7 @@ import { ListItem } from 'material-ui/List';
 import CircularProgress from 'material-ui/CircularProgress';
 import Avatar from 'material-ui/Avatar';
 import ActionAssignment from 'material-ui/svg-icons/action/assignment';
-import { standardDateFormat, shortTimeFormat } from 'lib/date_time_defaults';
+import { standardDateFormat, shortTimeFormat } from 'lib/dateTimeDefaults';
 
 const iconStyle = {
   width: 32,
