+++ client/app/bundles/course/enrol_requests.js
@@ -1,3 +1,3 @@
-import initializeAjaxForms from 'lib/helpers/table_row_form_helpers';
+import initializeAjaxForms from 'lib/helpers/tableRowFormHelpers';
 
 initializeAjaxForms('tr.course-enrol-request .approve');
