+++ client/app/bundles/course/experience_points_records.js
@@ -1,3 +1,3 @@
-import initializeAjaxForms from 'lib/helpers/table_row_form_helpers';
+import initializeAjaxForms from 'lib/helpers/tableRowFormHelpers';
 
 initializeAjaxForms('tr.experience-points-record #update');
