+++ client/app/bundles/course/users.js
@@ -1,3 +1,3 @@
-import initializeAjaxForms from 'lib/helpers/table_row_form_helpers';
+import initializeAjaxForms from 'lib/helpers/tableRowFormHelpers';
 
 initializeAjaxForms('tr.course-user #update');
